fix Image, comments, qrcode

+++ src/repository/pictures.py
@@ -1,8 +1,7 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from sqlalchemy import select
-from sqlalchemy.orm import selectinload
-from sqlalchemy.orm import aliased
+
 from sqlalchemy import and_, desc
 
 
@@ -14,12 +13,12 @@ from src.database.models import User, Image, Tag, TagsImages
 
 
 from src.services.cloud_image import CloudImage
+from src.schemas_pictures import ImageModel
 from cloudinary import CloudinaryImage
 import qrcode
 
 
 def create_taglist(tags: str) -> list:
-    print('create_taglist')
     return [tg for tg in tags.strip().split(' ') if '#' in tg][:5]
 
 
@@ -27,12 +26,11 @@ def create_taglist(tags: str) -> list:
 
 
 async def add_tags_to_db(tags: str, image, db):
-    print('add_tags_to_db')
     tag_list = create_taglist(tags)
     for tg in tag_list:
         psql = select(Tag).filter_by(tag=tg)
         result = await db.execute(psql) 
-        if not result.fetchone(): #db.query(Tag).filter_by(tag=tg).first()
+        if not result.fetchone(): 
             new_tag = Tag(tag=tg)
             db.add(new_tag)
             await db.commit()
@@ -40,7 +38,7 @@ async def add_tags_to_db(tags: str, image, db):
         else:
             psql = select(Tag).filter_by(tag=tg)
             result = await db.execute(psql)
-            new_tag = result.fetchone() #db.query(Tag).filter(Tag.tag == tg).first()
+            new_tag = result.fetchone() 
         tag_pic = TagsImages(image_id=image.id, tag_id=new_tag.id)
         db.add(tag_pic)
         await db.commit()
@@ -53,10 +51,9 @@ async def create(description: str, tags, image_url: str, public_id: str, user: U
     :param description: str: The description of the image
     :param image_url: str: The url of the image
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     """
-    print('create')
     image = Image(description=description, image_url=image_url, public_id=public_id, user_id=user.id)
     db.add(image)
     await db.commit()
@@ -74,24 +71,16 @@ async def get_images(limit: int, offset: int, user: User, db: AsyncSession):
     :param limit: int: The number of images to return
     :param offset: int: The number of images to skip
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A list of image objects
     '''
-    # images = db.query(Image).filter(and_(Image.user_id == user.id)). \
-    #     order_by(desc(Image.created_at)).limit(limit).offset(offset).all()
-    print('get_images')
-    # psql = select(Image).filter(and_(Image.user_id == user.id)). \
-    #     order_by(desc(Image.created_at)).limit(limit).offset(offset)
-    # print(psql)
-    # result = await db.execute(psql)
-    # images = result.fetchall()
-
-    
+   
     result = await db.execute(select(Image).filter(Image.user_id == user.id).order_by(Image.created_at.desc()).limit(limit).offset(offset))
-    print(result)
     images = result.fetchall()
-    print(images)
-    return images
+    im_dick = []
+    for o in images:
+        im_dick.append(o[0])
+    return im_dick
 
 
 
@@ -101,17 +90,16 @@ async def get_image(image_id: int, user: User, db: AsyncSession):
     
     :param image_id: int: The id of the image to return
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('get_image')
-    # image = db.query(Image).filter(and_(Image.user_id == user.id, Image.id == image_id)). \
-    #     order_by(desc(Image.created_at)).first()
-    psql = select(Image).filter(and_(Image.user_id == user.id, Image.id == image_id)). \
-        order_by(desc(Image.created_at))
-    result = await db.execute(psql)
-    image = result.fetchone()
-    return image
+
+    psql = await db.execute(select(Image).filter(Image.user_id == user.id, Image.id == image_id))
+    image_data = psql.fetchone()
+    return image_data[0]
+
+    
+
 
 
 async def get_image_from_id(image_id: int, user: User, db: AsyncSession):
@@ -120,16 +108,14 @@ async def get_image_from_id(image_id: int, user: User, db: AsyncSession):
     
     :param image_id: int: The id of the image to return
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('get_image_from_id')
-    #image = db.query(Image).filter(and_(Image.id == image_id, Image.user_id == user.id)).first()
+    
+    psql = await db.execute(select(Image).filter(Image.id == image_id, Image.user_id == user.id)) 
+    image = psql.fetchone()
 
-    psql = select(Image).filter(and_(Image.id == image_id, Image.user_id == user.id))
-    result = await db.execute(psql)
-    image = result.fetchone()
-    return image
+    return image[0]
 
 
 async def get_image_from_url(image_url: str, user: User, db: AsyncSession):
@@ -138,12 +124,11 @@ async def get_image_from_url(image_url: str, user: User, db: AsyncSession):
     
     :param image_url: str: The url of the image to return
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('get_image_from_url')
-    #image = db.query(Image).filter(and_(Image.image_url == image_url, Image.user_id == user.id)).first()
-    psql = select(Image).filter(and_(Image.image_url == image_url, Image.user_id == user.id))
+
+    psql = select(Image).filter(Image.image_url == image_url, Image.user_id == user.id)
     result = await db.execute(psql)
     image = result.fetchone()
     return image
@@ -155,14 +140,16 @@ async def remove(image_id: int, user: User, db: AsyncSession):
         
     :param image_id: int: The id of the image to delete
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('remove')
     image = await get_image_from_id(image_id, user, db)
-    db.delete(image)
-    await db.commit()
-    return image
+    if image:
+        await db.delete(image)
+        await db.commit()
+        return image
+    else:
+        return None
 
 
 async def image_editor(image_id: int,
@@ -175,19 +162,19 @@ async def image_editor(image_id: int,
     :param image_id: int: The id of the image to edit
     :param body: EditImageModel: The body of the request
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('image_editor')
+
     image = await get_image_from_id(image_id, user, db)
     if image:
         edit_data = []
+        
         if body.circle.use_filter and body.circle.height and body.circle.width:
             trans_list = [{'gravity': "face", 'height': f"{body.circle.height}", 'width': f"{body.circle.width}",
                            'crop': "thumb"},
                           {'radius': "max"}]
             [edit_data.append(elem) for elem in trans_list]
-
         if body.effect.use_filter:
             effect = ""
             if body.effect.art_audrey:
@@ -200,7 +187,6 @@ async def image_editor(image_id: int,
                 effect = "cartoonify"
             if effect:
                 edit_data.append({"effect": f"{effect}"})
-
         if body.resize.use_filter and body.resize.height and body.resize.height:
             crop = ""
             if body.resize.crop:
@@ -211,12 +197,10 @@ async def image_editor(image_id: int,
                 trans_list = [{"gravity": "auto", 'height': f"{body.resize.height}", 'width': f"{body.resize.width}",
                                'crop': f"{crop}"}]
                 [edit_data.append(elem) for elem in trans_list]
-
         if body.rotate.use_filter and body.rotate.width and body.rotate.degree:
             trans_list = [{'width': f"{body.rotate.width}", 'crop': "scale"}, {'angle': "vflip"},
                           {'angle': f"{body.rotate.degree}"}]
             [edit_data.append(elem) for elem in trans_list]
-
         if edit_data:
             CloudImage()
             new_image = CloudinaryImage(image.public_id).image(transformation=edit_data)
@@ -236,10 +220,9 @@ async def edit_description(image_id: int,
     :param image_id: int: The id of the image to edit
     :param body: EditDescriptionModel: The body of the request
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('edit_description')
     image = await get_image_from_id(image_id, user, db)
     if image:
         image.description = description
@@ -255,10 +238,9 @@ async def qr_code_generator(image_id: int,
     
     :param image_id: int: The id of the image to edit
     :param user: User: The user object
-    :param db: Session: A connection to our Postgres SQL database.
+    :param db: AsyncSession: A connection to our Postgres SQL database.
     :return: A image object
     '''
-    print('qr_code_generator')
     image = await get_image_from_id(image_id, user, db)
     if image:
         image_url = image.image_url
@@ -270,8 +252,8 @@ async def qr_code_generator(image_id: int,
         qr.add_data(image_url)
         qr.make(fit=True)
         img = qr.make_image(fill='black', back_color='white')
-        img.save(f'./src/services/qr_codes/{image_id}.png')
-        image.qr_code_url = f'./src/services/qr_codes/{image_id}.png'
+        img.save(f'./src/services/qrcodes/{image_id}.png')
+        image.qr_code_url = f'./src/services/qrcodes/{image_id}.png'
         await db.commit()
         await db.refresh(image)
         return image
